build: rename dagger module following official conventions
