+++ .dagger/lint.go
@@ -3,8 +3,9 @@ package main
 import (
 	"context"
 
-	"github.com/openmeterio/openmeter/dagger/internal/dagger"
 	"github.com/sourcegraph/conc/pool"
+
+	"github.com/openmeterio/openmeter/.dagger/internal/dagger"
 )
 
 func (m *Openmeter) Lint() *Lint {
